+++ src/libs/blueprint/conduit_blueprint.hpp
@@ -17,6 +17,7 @@
 #include "conduit.hpp"
 
 #include "conduit_blueprint_exports.h"
+#include "conduit_blueprint_config.h"
 
 #include "conduit_blueprint_mesh.hpp"
 #include "conduit_blueprint_mesh_examples.hpp"