+++ CHANGELOG.md
@@ -14,8 +14,9 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 
 
 #### Blueprint
-- Added `blueprint::mesh::examples::related_boundary`, which creates a multi-domain mesh with a related boundary topology and several fileds that encode relationships between the main topology and the boundary.
 - Added `blueprint::mesh::examples::polystar`, which creates a mesh with a polyhedral star pattern that demonstrates hanging vertices in a topology, along with its representations from `generate_sides` and `generate_corners`.
+- Added `blueprint::mesh::examples::related_boundary`, which creates a mulit-domain mesh with a related boundary topology and several fileds that encode relationships between the main topology and the boundary.
+- Expanded `blueprint::mpi::mesh::generate_partition_field` to support all topology types.
 
 
 ### Fixed